Rename the abstraction for stat procc enchants
