+++ src/parser/retail/modules/features/enchants/dragonflight/FrozenWrit.tsx
@@ -2,7 +2,7 @@ import ITEMS from 'common/ITEMS/dragonflight/enchants';
 import SPELLS from 'common/SPELLS/dragonflight/enchants';
 import { Options } from 'parser/core/Analyzer';
 import STAT from 'parser/shared/modules/features/STAT';
-import Writ, { SECONDARY_STAT_WRIT_VALUES } from './Writ';
+import StatProccEnchant, { SECONDARY_STAT_WRIT_VALUES } from './StatProccEnchant';
 
 // ================ SAMPLE LOGS ================
 // Frozen Writ R1
@@ -27,7 +27,7 @@ const RANKS = [
   },
 ];
 
-class FrozenWrit extends Writ {
+class FrozenWrit extends StatProccEnchant {
   constructor(options: Options) {
     super(STAT.VERSATILITY, SPELLS.FROZEN_WRIT_BUFF, RANKS, options);
   }
